Improve and simplify validation errors
If we are precise in the context, it should be easier to understand the
validation errors. Make them more explicit and succinct, which should
help the reader when an error occurs.

While here, improve the notion in the case where an interface check is
failed. It printed the whole internal representation of the interface
which was very spammy. This can be avoided.