Do not wrap types in {ty, •}
Whenever we have a field in the parser which is a type field, it is in a context
where we know it is a type. Hence, there is no reason to wrap the Type in
a wrapper that tells you it is a type. The rest of the code simply runs hoops
to unwrap these {ty, •} blocks whenever it needed the underlying type.

Fix this by never creating the wrapper. Once we have no wrapper, we can
go ahead and remove some ugly code from the system.