Do note try to derive types twice
The {scalar, binary()} notation is wrong. There is no way you can produce this
result in the code base at the parser level. Either you have one of the built-ins
in which case the known type is {scalar, builtin()} or you have an honest-to-god
real type, but it can't be known at the point of the code base where you are
elaborating. The actual details are to be resolved by the type checker.