Remove dead code
Every path to the input object check is guarded in a way such that it
can never be anything but an input object. Thus we can just assert that
we really have an input object at hand and crash otherwise. There should
be no way to input something here which is not an input object type:

* Either it comes from the elaborator, and it will never make the mistake.
* Or, it comes from the parameter checker, and it will neither make the
  mistake.