Fix handling of field args
Two things are added in this test case:

* A test for (optional) floating values in the dungeon by use of a
  plushFactor on monsters.
* Handling of args on fields in the right way.

When you type check a query document, you have the choice of what to run
your recursion. One option is to use the args in a field, the other is
to use the schemas knowledge of args in a field. The latter is the right
solution, as it allows you to grab default args and correctly non-null
args as well.

Once you add this change, the color modules resolvers turn out to be
wrong. They can't resolve correctly, because now fields will be there,
but with values of `null`.