Handle list types correctly.
When we have written ['Property'] in the schema definition, canonicalize
this into {list, 'Property'} and use this canonicalization all over the place.

This makes the type checker able to correctly handle list types, something
which it couldn't before.

While here, make sure an Albino Hobgoblin which has DRAGON and
MURLOC genes work as intended.