Represent parsed list types as lists
The old code did a wrapping {list, L} when is_list(L), whereas the new code
just stores the list when we have a value. This makes the type checker able
to recognize lists of things as lists of things when they come as expressions
and this fixes a bug in GraphQL.

While here, dump a regression test case on the thing so it doesn't reoccur.