Fix input_object resolution
If providing an input object as part of a list, and as an expression, the
problem was that we end up with an object containing
such objects into the right form before processing them.

To make sure this doesn't happen again, make it possible to add multiple
stats to a Monster. In principle this would mean we should add a Two-Headed
Ogre, but skip that for now.