Implement complete_value/6
This function follows the spec more or less precisely. We then take
the output of this function and graft it onto our normal execution stack.

The function implements "complex values" correctly and provides the
right output in every case pertaining to processing of complex values.

Still missing is proper error handling. The spec is not very precise about
how errors propagate in the system, so we will have to handle this in a
later commit.