Test for and validate enumeration values
When we try to look an enum value, it may not exist. When this happens,
we ought to throw a correct error message for the query and push that
to the client. While here, test a couple of different invalid invocations
to make sure we nail errors down and make them unlikely to occur.