Simplify the elaboration and creation of var defs
The code did the same thing twice. First it elaborated the types of vardefs
and then it stuffed those into a `schema` field on each vardef. Next, when we
created the actual varenv, we did the same code again.

Remove the `schema` field and stuff the types directly in `ty` with elaborated
values. This simplifies the type check as it can now directly utilize the types
we elaborated for the var env.