Remove the somewhat dubious schema_obj entry from #field{}
Just use a more low level solution based on adding the type inside the
away and the `schema` record-field can now handle all of what we need
in order to make things work.

Things are a bit hacky in the execution engine as a result of this change,
but there are good indications that we can simplify the execution later
on.