Simplify collect_fields/6
We can simplify a lot of the internals in the collect_fields/6 function by moving
more stuff to the top level pattern match and reuse paths through recursion.

If we get a named fragment, we can just look it up and then tail-call into
ourselves as if it was an anonymous fragment. This simplifies the code base
and removes a code path which was the same anyway. While here, we can move
more stuff to the top-level pattern match.