Remove {non_null, ...} hack in list completion
The problem with the list completion has to do with the recursion. The
recursion has to take into account the fact that if an inner value
returns an error case, it must factor through the non-null case in
complete-value, which it did not. With this change, the factoring will
happen correctly, even if the value is {error, Reason}:

* The code will run the non-null wrapper
* The code will then turn it into a proper error if needed

The solution is to rearrange the complete-value cases and handle them
accordingly. More stuff can be moved top-level, but this follows in a
later commit.