Track and inherit the surrounding object type in elaboration.
If you are given a query which has an inline fragment of the form

... {
  x
  y
}

The rule is that this is a fragment on the same type as the thing
which surrounds that inline-fragment, e.g. in a context of type
Monster, the fragment would be like having written

... on Monster {
  x
  y
}

This is especially important in the case of using directives on such
objects:

... @include(if: $fat) {
  x
  y
}

in which case the include statement triggers and decides upon what
this piece of code actually means.

The patch here, augments fragment to take an optional OType which is
the type context. If the inline-fragment itself has no given type
condition, we just inherit the type condition from the given OType. If
we have a named fragment, it won't have an OType, but then it MUST
have a type condition, so we are still safe.