Fix list processing in the type checker
Here is the problem:

When we want to type check a list of things, we generate the value
types for each element in the list, and then we proceed to check for
reflexivity on the list. The problem is that once we have processed
the whole list, some of the values inside the list may be changed
through a replacer. So we replace the whole list. But this sometimes
injects a type in place of a value.

Fix:

When generating list value types, generate pairs, {T, V} where T is a
type and V is the original value. When you try to reflect the list,
use the replacer to figure out if you need to use the original value
or the replaced one. This makes the whole-list-replace able to run
without injecting types.