Abort computation if the value type doesn't coerce correctly
If we hit a case where the value we are given doesn't match the type
we have looked up, we can't coerce the value type to what is given.
This is always a bug in the input, so we can just abort the
computation rather than returning a function clause bug.