Remove the dependency on lager
We replace lager with the built-in error logger. It strips off the
lager dependency and we don't need that since the error logger can do
its job. Furthermore, lager will pick up the error logger messages
anyway and transform them in the system, so we don't lose anything by
doing this.

The change brings the dependencies from 1 to 0.

Reported-By: @goertzenator -- who noted that GraphQL is 0.38megabyte
in size whereas lager is close to 2megabyte.