Properly handle errors on type resolution.
When you resolve an abstract type to a concrete one with the type
resolver, it can either crash or return `{error, Reason}` for some
reason given by the developer. This was not properly handled in the
execution phase.

Introduce a helper for abstract type resolution and use it in the case
of unions and interfaces. Once present, return errors upon errors.

While here, provide a `kraken` object in the dungeon scheme which is
nonsensical and use this to force the system into the error path,
testing that the error is indeed detected by the system.