Coalesce error path reversal to one place
When the execution function errors, there should be one place where
the path component is reversed. Make that be the err/2 function.

Alter the test cases now errors are reported correctly for every kind
of error which can occur. This eliminates some nonsensical errors as
well and locks down proper error handling a bit better.