Improve error messages when a resolver crashes
There is no reason to return the full internal object type from a
resolver crash. Rather, return what object type and what field name
crashed to the outside world, but keep the full information about the
error message internally in the error logger. This way, the output is
a bit easier to read, but the you still have the full error internally
to work with.