Fix GetOperation for parameterized operation calls
The variant where your query has a single operation with parameters
was not handled correctly. It was rejected, but the spec says it
should be accepted.

The problem is in the type checking phase for parameters. It didn't
account for this particular case correctly, so you obtained the empty
map as the parameters list. This of course breaks deep down in the
execution code since it assumes all parameters have been correctly
checked.

The fix is to analyze the environment and if there is a single
operation, with no "OpName" given, then assume it is *the* operation
to execute.