Capture errors on value completion of array/list types
When an array completes, we expect the result to be of the type
[{ok, _} | {error, _}]. If the programmer inputs something else,
exit resolution and print out an error. This is a bug and it needs
fixing in the code.

Also, output an error message which the programmer can use to figure
out where the error is.

We've seen this error crop up internally at ShopGun many times, so
we've decided to do something about it in the GraphQL engine itself:
assert the return is valid.