Streamline error handling in the execution phase
Make the execution phase use the same error handler as the earlier
phases which streamlines error handling by quite a lot. This patch
also paves the way for developer-controlled error handling where an
error in a resolver function gets passed to a module responsible for
handling that kind of error.

Resolution of lists still has a bug where multiple errors on the same
path occurs, but we should be able to handle that in a later patch as
well. The bug was not introduced as part of this patch, but has
already been there, so we ignore it for now.