Improve the correctness of graphql_execute:value/2
Rewrite the function to use a type-guided operation. We case-analyze
on the type of the incoming value first. Then, once the type is known,
we evaluate the possible value for validity. This ensures errors occur
at the right spot.

The patch paves the way for better enumeration handling as well as
better handling of values in general when resolving args for a
function.