Refactoring of the dungeon test suite
Some of the tests in the dungeon suite rely on some known monsters and
rooms to function, especially the ones that test static input values
in a graphql query document. I've solved this by having a range of
item ids will be created in a deterministic order during test setup,
the id counter is then fast forwarded to 1000 and from then on rooms
and monsters created in tests will be in that range.

This will allow us to have some known monsters and rooms without the
static test data will get messed up by dungeon and room generation.

I've added some helpers to support test data generation, one that
create monsters and one that deals with rooms; these functions take a
configuration option so the monster or room can be created with a
given amount of hit points, name, etc. There is also support for
skipping a number so one can get an id that is guaranteed to not
exist.

This commit also deal with some trailing whitespace.