Support 3-ok-tuples
In some cases it is convenient to pass auxiliary data with the result;
data that should not go directly to the end-user but should trigger
some side-effect in the transport before sending data/error to the
response body. This has been solved by using message passing and
collect these just before we complete the top-level.

In resolvers one can, besides the two-ok-tuple (which has the old
behaviour), return a three-ok-tuple which behave like this:

```
{ok, any(), [term()]}
```

The response map will now contain an `aux`-field which contain a list
of terms, which can be used to modify the response object, such as a
Cowboy response object, and set a cookie, add a header, etc.

N.B.: Errors cannot set auxiliary data, and one should not rely on the
order in the auxiliary data list.