Initial parallelization of GraphQL queries
Ready a proxy process for deferred execution. If a GraphQL object
requires deferred execution, use the proxy process to handle data
routing for the subquery and ready a new deferral process for the next
layer in the recursion.

There are still loose ends, but it looks like the basis are there.