Coalesce all the wait-state into a single #workunit{}
Introduce a single #workunit{} record which maps everything into it.
Then start refactoring around this new construction. It greatly
simplifies the flow-path in the code since there is now only a
singular flow which needs to be maintained in the system.

While here, get rid of the 'routes' entry in the defer state since an
object itself can contain what given 'target' it wants its result sent
to later on.