Add some more nastiness to the tests
This forces the system to create some more deferral states which in
turn messes up the system some more. Things are not looking like they
should again, but progress is being made in understanding how one
should proceed.

Current problems recurs around the idea of proper null-propagation in
the Graph when errors occur, and the fact that deferrals can push
value completion ahead of it. This is not handled correctly right now.