Implement a closure-based system for doing deferrals
Tidy up the closure based system. Do not pass names around and keep
explicit track of missing values. This improves the current state of
the system and simplifies the internals quite a bit. Makes every case
but the nastier list cases pass in the system (which is not currently
surprising: lists aren't handled yet in this system/model)