Use records over tuples for passing values
This change introduces passing via records over tuples. It is done in
anticipation of a later change where we need to attach more data into
the data passed around in the graph. With records, we can just add
that data later on as needed, whereas tuples requires lots of changes
all over the code.

While here, fix the remaining problems with nested closure invocation.