Forward closures via field-names/indexes over references
The old way of forwarding data in the graphql answer was by using the
reference of a closure as the key. The problem with this solution is
that when a field is deferred and then deferred again, the reference
changes and the original reference is unusable. The earlier solution
was to introduce a special forwarder-closure in this case, but this is
not really needed.

Alter the code such that an `Upstream` is now `{UpstreamRef, Key/Index}`
where key is either a field-name or an array-positional index.
Whenever we recurse, we know exactly what the Key/Index should be.
Then arrange that forwards happen on the Key or Index rather than the
closure reference (which is not stable due to the above observation).

The top level introduces a special key `top_key` as an atom to handle
its case. Apart from that, the top level works exactly like everything
else.

Now, a #done{} record can enforce that there is *always* a key to
forward to. This helps consistency and lets the dialyzer work on
correctness for us.

We can also kill the forwarder-closure in the code base as it is not
needed anymore(!)

While here, clean up types in the records to make the dialyzer more
strict.

End result: more consistent code.