Improve routing code, revert key-passing
In a recent commit, we changed the way inner closures pass data up the
response tree. Rather than passing `Upstream`, we passed `Upstream,
Key` so underlying elements in the tree new what key they "sat under".
This was deemed necessary, because the closure itself might be
replaced by another closure if you have a defer sitting under another
defer.

However, this runs into trouble once you start considering the concept
of cancellation. If an object or array errors, we want to cancel every
immedate sibling. But if the only thing the object knows about are the
keys, then we don't know whom the siblings are and we can't message
them. This suggests we should keep references to our children, but
this runs into the problem above.

The solution is to have a way to change the route through a
{change_ref, From, To} message to a closure. This will update the
closure with the new route. We do this *before* we read the next
message off of the mailbox, so we don't violate ordering problems.
With this in place, we can still cancel siblings, and we can run defer
in defer.