Track canceled targets
When we process cancellations, we now track which targets we
canceled. In the event where there is a response to something we threw
away, we now skip that particular message without sending a warning in
the log. This will shut up the log by quite a lot and let us handle
the notion of canceling messages.