Do not allow a *set* of worker pids for a field defer
Here is a problem: suppose that you are watching/monitoring a set of 2
or more worker pids. Suppose they start failing. Do you treat the
failure as logical AND or as logical OR: should all or just one fail
before it is a total failure? How should you treat cancellation upon
crashes?

In the unary case, this doesn't matter.

Change the API so there is only a single worker. If a user needs
multiple workers, the solution is obviously to let the user handle
that themselves by means of a helper which can handle the monitoring
going on. It is also far simpler to document.