Remove the default resolver from the code base
The change in the execution engine is straightforward.

However, we need to give the introspection module a resolver execute/4
function for handling its standard cases. Otherwise things starts
failing in the system.

While here, go through test cases and remove their reliance of the
default resolver. Insert explicit resolvers by factoring back to the
modules themselves through the `resolve_module` entity, which is the
correct way to do it.