When parsing GraphQL, use builtin types
Any parse can lookup built-in types in the schema. Do that. This
eliminates the need for "Magic" {scalar, TyAtom} entries in the code
base and simplifies later stages. Any type in the system is now of the
form #scalar_type{}, user defined or not. This allows us to remove
certain code paths in the future since they don't have to be there
anymore.

While here, Cuddle with the star_wars schema. It injects via dynamic
injection and now a string is given as 'String'. This is more precise
and less error-prone anyway.