Remove any mention of {scalar, _} in the code base
Since we are now elaborating built-in types as user-defined types, we
can go ahead and remove lots of code in the system which purports to
support this notion.

Some tests which uses the dynamic injection code needs to be updated
since 'bool' isn't a valid type anymore. You are to use the builtin
'Bool'.

Speaking of bool, it was never handled correctly as a type in the code
base, so make sure it is now handled like other built-in types.