Remove the need for "{replace, X}"
In the new world order we don't need to handle replacements specially
since everything is really a replacement. This simplifies parts of the
code because we can just handle the replacements directly as they
occur.

A later patch will unwind some of the error-passing and use exceptions
instead. It will ensure the error handling is truly a side-channel in
the system as you would expect.