Rewrite the type checker
The old type checker, that this patch replaces, worked by analyzing a
value for its possible type through inference. And then it tried to
match the inferred type with the schema type through a call `refl`.
This meant the type checker could generate a new "world" so to speak
of the Value Types, which is then used for its judgment.

The new type checker is cleaner than the older one. It runs by two
inductions, the first over the values and the second over the schema
type. Each case analyzes what is valid in that case. This makes the
type checker simpler to follow as when we case split over the value
first, we now have a smaller context, which makes it easier to
understand what the valid schema types are.

There are two major functions in the new type checker:

type_dec(DTy, STy) is a decider. It decides if DTy is a valid
type (from the query document) when we expect an STy (a type from the
schema). It is used to handle type-decisions on variables (that is, if
your document contains parameters).

judge(Ctx, Path, DValue, SType) is the typing judgment. It generally
follows induction over the value expression by case splitting the
value into its different cases. For literal values, the judgment
coerces inner induction: case splitting over the claimed type in the
schema. Here `DValue` is the document value, and `SType` is the schema
type.

To correct enumerations, the way enums are resolved has been updated
such that the function takes a string as input. This handled both the
case where an enum is part of a query document (and given as a special
enum type) and part of a JSON input (in which case the enum is coerced
from a string literal).

Correct two test cases: the old type checker failed to provide the
correct path in some type errors.

Add documentation all over the place.