Coerce default values as well
This plugs a hole in the coercion code with enumerated types. If an
enum is supplied as a default it was never coerced, but because the
default value was {enum, X} and the default coercer returned {enum, X}
nobody ever saw this.

Default values are now param-checked like everything else. If they
fail, the output error is different and we drop an error in the error
logger because this is a misconfiguration.