Handle the coercion of scalars into lists
The case where you have e.g., a `Pet` in a context which expects a
`[Pet]` is valid. The way this is handled is that coercion happens and
the value of the Pet, say `V`, is put into a single list `[V]` before
being handled. This ensures that clients can be lazy and input single
elements in list contexts and everything will work out.

The solution here is to postpone the coercion into the
execution/runtime phase. It has to be at this spot because we want to
add fragment variable handling later and this can only work if the
coercion all happens at the execution phase. Otherwise you run into
trouble.

We first type check the coercion in the statics phase of the system.
If it type checks, then we know we can safely do a coercion in the
execution phase without having to resort to a lot of checking. Rather,
we can rely on the fact that the type checker already passed the code,
so the type embedding we are looking at must be correct.

Enable the appropriate tests for this while here.