Handle validation 5.5.1. This closes #113
The resolution of this validation is a bit more involved.

Amend the validation suite with a field on which we can check for
input object uniqueness. This initially fails since we don't correctly
check for it.

The parser creates a map out of input objects. If we are to report the
error, then it is easier to hand a tagged input object to the rest of
the code base, and then handle it there.

The type checker can now handle the case where it is given an
input_object and then check for uniqueness. If the test passes, we
create the map so the rest of the code can use the expected correct
type. Failure is turned into error.

Handle the corner case of input arguments as JSON data. These will
already be maps and it is *undefined* in this case what happens if a
key occurs twice in JSON. It depends on the parser which is employed.
We temporarily turn such a list into a map for further processing
through the normal pipeline of events.