Handle validation 5.6.2. This closes #115.
In order to handle this we proceed by doing a number of things to the
system.

First, we remove hardcoded handling of directive locations in the code
base altogether.

Second, directives now has a new field, locations, which encodes where
those directives are valid.

Third, the elaboration step for directives is now extended
to take a 'context' from where we came when checking a directive. We
can use this context to verify that the directive is valid in the
given location. A simple list membership is enough.

We can then remove some special cases in the code and replace
everything with default paths which are then correctly traversed
instead.

While here, be proactive and open up the work for supporting 5.6.3 as
well: add a uniqueness checker. This moves the checker itself to
graphql_ast and also simplifies its API: it now sorts itself.