Attach the basis for a fragment embed judgment
In order to correctly check for the validity of a fragment spread in a
given scope, we add a new judgment form fragment_embed. This form is
currently not capturing anything, but we inject it in the right places
in the code base and arrange that the types for fragments indeed
produces valid types as input.