Implement support for looking up interface implementors
In order to verify that interfaces can be used correctly in fragment
spreads, we need a way to look up the implementors of an interface.
This information is not readily present, to scan the set of objects
for it for now through the use of QLC. In a later solution, we can
carry a 2nd index on the schemas OBJECTS table to speed this up, but
for now, use a simple solution which will work.