Fix order in the main type-checking judgment
We don't have that the different values and types are disjoint, so
there is an overlap in the input coverage. In this setting, variables
should be judged *before* we judge nullability. This makes sure the
type checker does the right thing in this case, and rejects a document
which is not valid due to input non-null checking.