Thread the VarEnv into param checking
Since you can have a parameter check which involves a variable lookup,
we need to thread the VarEnv into the param checker code. While here,
move the on-demand elaboration code to the top of the function. This
ensure that the code always first elaborates, if it hasn't been done,
and then it does everything else. Variable Type checking is one of
these things.