Fix #168: Do explicitly fail on already existing entries
If we try to insert an Entry into the GraphQL schema, but such an
entry is already present in the schema, then exit the process with a
failure. These are not allowed and should result in a hard crash so
the programmer knows that something needs fixing the schema
specification.

This handles the clarification that you are not allowed to enter the
same specification twice.