Use the dialyzer to strengthen data flow
The dialyzer has reported a number of errors in the data flow in this
area of the code base. Proceed by correcting the data flow and make it
more explicit, returning ok | {error, Reason} everywhere. This solves
the dialyzer warnings, and it makes it more explicit what is going on.

In addition is corrects a number of type specifications which were off.