Fix #159 - Support unnamed queries more leniently
We now support a query of the form

    query (...) { ... }

where there is no name for the query. This follows general rules for
not supplying a name:

* It *must* be the sole query in the document.
* Things must still match up

While here, also accept "" as an operationName as define it as a
short-hand for the unnamed query. Again, a lot of clients (Javascript,
I am looking at you!) supply "" as a synonym for null so to make life
easier on their side, cast "" into the undefined query.

This fixes #159.

Also, while here, provide test cases for the code paths to make sure
the new way of calling works as expected.