Allow the output of floats in integer contexts
If we have a float value such as 1.0 it is truncated into 1. This
allows the GraphQL server to output that value in the context under
the guise of being an integer. Reject all other floats.

This might be useful in Erlang since there are numerous examples where
integers are converted into numbers.