Lookup of a field type cannot be null.
The Jun2018 spec has explicitly written that a field type lookup
cannot be null. Indeed, the code agrees on this and will never return
the value null for a field type lookup anymore. We have changed the
value of the lookup to "not_found" anyway., so remove the need for
having a null lookup.