Fix a not_found field lookup. Closes #173
The Jun2018 Spec explicitly says that this is a case which cannot
happen. Indeed, every field type in execution must be present for a
lookup. Otherwise, the type checker is wrong. Remove the clause and
force a maps:get/2 crash if that isn't the case.

This solves yet another Jun2018 spec case.