Fix null-input handling
In anticipation of Jun2018, we must eventually allow an input of a
null. The older parser accepted nulls by lifting that keyword into an
identifier, but we cannot allow that anymore. Hence, we accept 'null'
as a GraphQL Value now in the parser.

This in turn means the type checker has to reject the situation. It
has a relatively easy task at hand because the argument picker
take_arg/3 can be changed to disallow reading a null-arg.

While here, add tests for this behavior in the case of non-null and
null arguments. This makes sure we lock it down as a regression.