Adding directive definitions and directive validation in schema
Missing is anything else directive definition related, specifically
directive checks in query documents and also the following directive locations in schema:
- ARGUMENT_DEFINITION
- ENUM_VALUE
- INPUT_FIELD_DEFINITION
are not yet validated.

I havnen't looked at if there are any special introspection things going on with
directives, so if that is that case, that is also missing.