Handle field resolver arity=3 in the canonicalization instead of in the execution
I was about to do directive introspection, but found this instead.

Seemed like a resolver with arity 3 was only ever used when a field's resolve is set.
This seems to mostly be a feature useful in the instrospection where you'd like to
overwrite a field resolver instead of using the resolve module on the object.
The blog example also inlines field resolvers, which makes sense for small applications.

However, it seemed like a convenience thing that should be handled in canonicalization
so I patched that so that a field resolver always has arity 4 in the execution.

- - -

I'd also like to unify the way the resolver is specified on objects and fields.
It's a bit weird that an object takes a `resolve_module` which must be a module,
and a field takes a `resolve` which must be a function.

Naming both `resolve` and accepting both an atom (which would call a module's execute)
or a function of arity 4 (or 3) seems to be more consistent, and makes it more obvious
that it's a list of possible resolvers, taking the field resolver if available, otherwise
using the object resolver.
That's a different patch though.