Heed the specification for error paths
The specification in Jun2018 says that we must have the path follow
the response documents keys. So in particular, the operation we
executed must not include the operation name itself. Hence, we remove
this part from the execution routine.

Also routine, we now remove that from test cases which failed.