Introduce extensions, kill Path from err-handlers
We introduce a genuine extensions map for the user to fill out. This
makes us spec compliant on this front for the future.

While here, remove the Path component from the error itself. As a
user, you shouldn't care about the path and you should never
manipulate it anyway, so amend the error with that structure later on.
Also while here, allow messages to be iodata() since it simplifies a
lot of stuff going forward.