Introduce an execution context in the execution phase
Remove `Path` from execution and push it into a context used for
execution called `#ectx{}`. This simplifies a lot of code in the same
vein as the simplification which were made in the type checker some
months ago.

We are not entirely done with the refactoring for now, but will
continue on this task at a later date.