Store schema as a persistent_term
The schema in a GraphQL engine has the property that it is very rarely
updated if at all, but is read into every process using GraphQL calls.
To avoid copying the term into every process, we store this as a
persistent_term upon schema validation.

This avoids the copy into the process and shares the term for every
process in the system. The major gain is not the avoidance of ETS
locking, but avoiding putting the schema on the heap. For large
queries, this doesn't matter, but smaller queries which touch lots of
the schema, this should lead to a considerably leaner process.