Implement map/2 in the non-defer case
Strategy: keep a chain of changes. Apply this chain when we have value
resolution. Do not cope with the chain dumping a defer for now, but it
is doable given the closure handler anyway. It has to work with the
possibility that the upstream ref changes anyway, so it should be
sound to make this change on top.