Fix wrong list handling in variable coercion
Special handling of variable coercion is necessary in GraphQL because
if the schema specifies a type [A] then it is valid to use the type A
in its place (for convenience reasons).

However, the code did not contain the correct congruence relations.
Coercion of [A] with [B] is the same as running a coercion of A with
B, hoping that works out.

Add the missing congruence rule, then also alter test cases so they
cover this particular problem. While here, reorder names to be
consistent with the type checker.