Remove coercion from the type checker phase
We can't coerce in the type checker if things are to be done
correctly. The problem is that we have to pick up different defaults
in some situations. Consider the following query:

    query Q($k : Int) {
      foo(x: $k, y: $k)
    }

Where we call it with no $k parameter. Then, the defaults of x and y
must be picked up according to the specification. However, if we have
the definition as:

    type Query {
      foo(x : Int = 5, y : Int = 3)
    }

Then this won't work currently in the Engine. The reason is that we
have to pick either 5 or 3 for $k and that expansion can only happen
once in the system.

The idea is to push coercion down into the execution phase, where it
really belongs, and then shove information from the type checker along
for the ride. This allows different types to be picked up, and also
makes the type checker somewhat more robust.