Clean up how parameters are checked
Rather than passing a not_found on from a failed maps:get/3, handle it
locally in the parameter checker. This avoids a pattern match later on
in the code, which then streamlines the whole thing into a more fluid
flow.

This refactoring is needed for the overarching idea of handling
parameter passing correctly in the system.