Do not have a separate check_param pass
The check_param judgement was equivalent to the check_value judgement
except for enum values. Introduce a `sub_context` of either being in
query or variable mode when checking values. Then use this to
discriminate between the enum values, while streamlining every other
flow to the be the same.

This kills a lot of type checker lines and removes a lot of redundancy
in the type checker.